Docker tuning (#183)
* remove plex claim args

* make docker prune optional

* add upgrade script

* test uses single = for string equality

+++ sietchumbu/create-radarr.sh
@@ -14,4 +14,7 @@ docker run \
   -v /home/multi:/multi \
   -v /home/docker/inbound/complete:/downloads \
   ${CONTAINER}
-docker system prune
+
+if [ "$1" = "--prune" ]; then
+  docker system prune
+fi
